requirements.txt
fastapi
uvicorn
PyGithub
openai
scikit-learn
pandas
numpy
python-dotenv
requests
radon
pytest
pyngrok
groq
